:art: separated settings page/component
